changes login to modal
